+++ lib/client.js
@@ -1,8 +1,7 @@
 /**
  * Module dependencies
  */
-var Connection = require('./connection')
-  , events = require('events')
+var events = require('events')
   , crypto = require('crypto')
   , Store = require('./store')
   , eos = require('end-of-stream')
