--- lib/connection.js
@@ -1,57 +1,0 @@
-
-var mqttPacket = require('mqtt-packet')
-  , Connection = mqttPacket.connection
-  , inherits   = require('inherits');
-
-function emitPacket(packet) {
-  this.emit(packet.cmd, packet);
-}
-
-function MQTTConnection(duplex, opts) {
-  if (!(this instanceof MQTTConnection)) {
-    return new MQTTConnection(duplex, opts);
-  }
-
-  Connection.call(this, duplex, opts);
-
-  this.stream = duplex;
-
-  this.on('data', emitPacket);
-
-  duplex.on('close', this.emit.bind(this, 'close'));
-}
-
-inherits(MQTTConnection, Connection);
-
-[ 'connect',
-  'connack',
-  'publish',
-  'puback',
-  'pubrec',
-  'pubrel',
-  'pubcomp',
-  'subscribe',
-  'suback',
-  'unsubscribe',
-  'unsuback',
-  'pingreq',
-  'pingresp',
-  'disconnect'].forEach(function(cmd) {
-    MQTTConnection.prototype[cmd] = function(opts, cb) {
-      opts = opts || {};
-      opts.cmd = cmd;
-      // flush the buffer if needed
-      // UGLY hack, we should listen for the 'drain' event
-      // and start writing again, but this works too
-      if (this.write(opts)) {
-        if (cb) {
-          cb();
-        }
-      } else if (cb) {
-        setImmediate(cb);
-      }
-    }
-  });
-
-
-module.exports = MQTTConnection;
