+++ lib/server.js
@@ -6,11 +6,10 @@ var fs = require('fs')
   , net = require('net')
   , tls = require('tls')
   , util = require('util')
-  , Connection = require('./connection');
+  , Connection = require('mqtt-connection');
 
 function setupConnection(duplex) {
   var connection = new Connection(duplex);
-  duplex.on('error', connection.emit.bind(connection, 'error'));
   this.emit('client', connection);
 }
